Add BaseAuthResponseDialog with UserTokenClient support
