+++ libraries/Microsoft.Bot.Builder.Dialogs.Adaptive.Teams/Schemas/Actions/Teams.SendMessagingExtensionConfigQuerySettingUrlResponse.schema
@@ -2,7 +2,7 @@
     "$schema": "https://schemas.botframework.com/schemas/component/v1.0/component.schema",
     "$role": "implements(Microsoft.IDialog)",
     "title": "Send query link response",
-    "description": "Query link requests have activity.name='composeExtension/queryLink'.",
+    "description": "Query link requests have activity.name='composeExtension/setting'.",
     "type": "object",
     "properties": {
         "id": {
@@ -41,8 +41,8 @@
     "policies": [
         {
             "type": "allowedTrigger",
-            "kinds": [ 
-                "Teams.OnMessagingExtensionConfigurationQuerySettingUrl" 
+            "kinds": [
+                "Teams.OnMessagingExtensionConfigurationQuerySettingUrl"
             ]
         }
     ]
